requirements.txt
aiodns
alembic
annotated-types
black
brotli
bs4
httpx
langchain
openai
orjson
pep8
pip-chill
psycopg2
py-cord
pydantic-core
pylint
pyright
python-dotenv
rich
ruff
sqlmodel
